tts-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
gTTS==2.5.1
pydub==0.25.1
python-multipart==0.0.6
python-dotenv==1.0.0
